Fixing Display Transaction Summary

+++ CockaCola2/Transaction.java
@@ -35,7 +35,7 @@ public class Transaction {
    *
    */
 
-  public double getTotalSales() {
+  public int getTotalSales() {
     return this.totalSales;
   }
 
@@ -64,7 +64,7 @@ public class Transaction {
   public long getTransactionDate() {
     return this.transactionDate;
   }
- 
+
   /**
    * This method displays the Transaction that has occured when dispensing an
    * item.
